move Diff() to diff.go
